- Added LivingSpace class for simplicity of comparing if objects are out of range or not of defined living space.

+++ Game.cpp
@@ -4,8 +4,8 @@
 
 #include "Game.h"
 
-constexpr unsigned int WINDOW_WIDTH{1900};
-constexpr unsigned int WINDOW_LENGTH{1200};
+constexpr float WINDOW_WIDTH{1900};
+constexpr float WINDOW_LENGTH{1200};
 constexpr std::string_view WINDOW_TITLE{"Game!"};
 
 Game::Game() :
